pyproject.toml
[project]
name = "scheduling"
version = "1.0.0"
description = "Teacher Outlook calendar synchronization for Swiss Learning Group"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=23.0",
    "tenacity>=8.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
]
